backend/requirements.txt
amqp==5.1.1
asgiref==3.5.1
autopep8==1.6.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.6
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.0.4
django-environ==0.8.1
django-extensions==3.1.5
django-filter==21.1
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.11.3
install==1.3.5
kombu==5.2.4
Markdown==3.3.7
mysqlclient==2.1.0
prompt-toolkit==3.0.29
pycodestyle==2.8.0
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
zipp==3.8.0
